pgml-dashboard/content/blog/benchmarks/hf_pinecone_vs_postgresml/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
datasets==2.14.4
dill==0.3.7
dnspython==2.4.2
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
loguru==0.7.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
pgml==0.8.1
pinecone-client==2.2.2
platformdirs==3.10.0
psycopg==3.1.10
psycopg-pool==3.1.7
pyarrow==12.0.1
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rich==13.5.2
six==1.16.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
xxhash==3.3.0
yarl==1.9.2
